+++ genomics_data_index/api/query/SamplesQuery.py
@@ -158,7 +158,7 @@ class SamplesQuery(abc.ABC):
 
     @abc.abstractmethod
     def _get_sample_names_query_infix_from_data(self, data: Union[str, List[str], pd.Series, SamplesQuery, SampleSet]
-                                            ) -> Tuple[Set[str], str]:
+                                                ) -> Tuple[Set[str], str]:
         pass
 
     @abc.abstractmethod