+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * Initial release of entire project.
 * Adding method to assign a score to clusters in a tree based on how well sets of samples are grouped together.
+* Changed up applying visual styles to sets of samples in a tree to only apply when rendering.
+    * This fixes an issue where copying trees was crashing when rendering large trees (I copy only once using 'newick').